Fix rss

+++ mee7/cogs/rss.py
@@ -6,34 +6,48 @@ from discord.ext import commands
 import feedparser
 import aiohttp
 
-
 class Rss(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
         self.bot.loop.create_task(self.update_feeds())
 
-    async def send_entry(self, entry, chn):
+    def format_html(self, text):
+        text = text.replace("<p>", "    ")
+        text = text.reaplce("<\\p>", "\n")
+        text = text.reaplce("<\\br>", "\n")
+        return text
+
+    async def send_entry(self, entry, chn: discord.TextChannel):
         embed = discord.Embed(
             title=entry['title'],
-            description=entry['summary'],
+            description=self.format_html(entry['summary']),
             url=entry['link'],
         )
+
+        if len(embed) > 1970:
+            embed = discord.Embed(
+                title=entry['title'],
+                description=entry['summary'][:1900]+"...",
+                url=entry['link'],
+            )
         await chn.send(embed=embed)
 
-    async def update_feed(self, feed_record = None, feed: feedparser.FeedParserDict = None, feed_id = None):
+    async def update_feed(self, feed_record = None, g_id = None, feed: feedparser.FeedParserDict = None, feed_id = None):
         if feed_record is not None:
             feed_id = feed_record['feed_id']
-            feed = await self.get_feed(feed_record['feed_url'])
+            feed = await self.get_feed(feed_record['feed_link'])
             if feed is None:
                 return
+
+            g_id = feed_record['g_id']
 
         chn_id = await self.bot.pg_con.fetchval(
             """
             SELECT rss_chn_id
                 FROM settings
                 WHERE g_id = $1
-            """, feed_record['g_id']
+            """, g_id
         )
         if chn_id is None:
             return
@@ -43,9 +57,9 @@ class Rss(commands.Cog):
 
         last_entries = await self.bot.pg_con.fetch(
             """
-            SELECT *
-                FROM rss_entries
-                WHERE feed_id = $1
+            SELECT feed_id, entry_id
+              FROM rss_entries
+             WHERE feed_id = $1
             """, feed_id
         )
 
@@ -65,12 +79,11 @@ class Rss(commands.Cog):
             )
         else:
             to_send = []
-            while True:
-                for entry in feed.entries:
-                    if entry['id'] in last_entries:
-                        break
-                    else:
-                        to_send.append(entry)
+            for entry in feed.entries:
+                if entry['id'] in [past_entry['entry_id'] for past_entry in last_entries]:
+                    break
+                else:
+                    to_send.append(entry)
             for entry in to_send:
                 await self.send_entry(entry, chn)
                 await asyncio.sleep(.25)
@@ -83,7 +96,9 @@ class Rss(commands.Cog):
                     """, feed_id
                 )
             else:
-                for past_entry in last_entries[::-1]:
+                for i, past_entry in enumerate(last_entries[::-1]):
+                    if i == len(to_send) - 1:
+                        break
                     await self.bot.pg_con.execute(
                         """
                         DELETE FROM rss_entries
@@ -98,12 +113,9 @@ class Rss(commands.Cog):
                 """, [(feed_id, e['id'], e['title'], e['summary'], e['link'], e['published']) for e in to_send]
             )
 
-
-
     async def update_feeds(self):
         await self.bot.wait_until_ready()
-
-        while self.bot.is_closed():
+        while not self.bot.is_closed():
             guilds_to_update = await self.bot.pg_con.fetch(
                 """
                 SELECT g_id, rss_chn_id
@@ -121,7 +133,7 @@ class Rss(commands.Cog):
                     SELECT *
                     FROM rss_feeds
                     WHERE g_id = $1
-                    """, guild_record['rss_chn_id']
+                    """, guild_record['g_id']
                 )
                 for feed_record in feed_records:
                     await self.update_feed(feed_record)
@@ -136,37 +148,60 @@ class Rss(commands.Cog):
                     return None
 
                 data = await resp.read()
-
         return feedparser.parse(data)
 
     @commands.command(brief="Add an rss feed. If no name is given, the title from the feed will be used")
+    @commands.has_permissions(administrator=True)
     async def add_rss(self, ctx, link, *, name=None):
         """add_rss [link] (name)"""
         for _ in range(2):
             feed = await self.get_feed(link)
-
             if feed is not None:
                 break
         else:
             await ctx.send("The url didn't work out, make sure it's correct and/or try again later")
             return
 
-        name = name or feed['title']
-
+        name = name or feed['feed']['title']
         feed_id = await self.bot.pg_con.fetchval(
             """
             INSERT INTO rss_feeds (g_id, feed_name, feed_link)
                  VALUES ($1, $2, $3)
-            ON CONFLICT (g_id, feed_name)
+            ON CONFLICT (g_id, feed_link)
             DO
                 UPDATE
-                   SET feed_link = EXCLUDED.feed_link
+                   SET feed_name = EXCLUDED.feed_name
             RETURNING feed_id
             """, ctx.guild.id, name, link
         )
-
         await ctx.send("If you have set an rss channel for your server, you will get updates for this feed starting.. now")
-        await self.update_feed(feed=feed, feed_id=feed_id)
+        await self.update_feed(g_id=ctx.guild.id, feed=feed, feed_id=feed_id)
+
+    @commands.command(brief="Remove an rss feed based on link")
+    @commands.has_permissions(administrator=True)
+    async def remove_rss(self, ctx, link):
+        """add_rss [link]"""
+        feed_id = await self.bot.pg_con.fetchval(
+            """
+            DELETE FROM rss_feeds
+                  WHERE g_id = $1
+                    AND feed_link = $2
+              RETURNING feed_id
+            """, ctx.guild.id, link
+        )
+
+        if feed_id is None:
+            await ctx.send("I didn't find a feed with that url")
+            return
+
+        await self.bot.pg_con.execute(
+            """
+            DELETE FROM rss_entries
+                    WHERE feed_id = $1
+            """, feed_id
+        )
+
+        await ctx.send("You won't be getting updates for that feed anymore")
